+++ lib/entities/tokens.dart
@@ -1,4 +1,3 @@
-import 'package:flutter_moshimoshi/core/utils/logs/hybrid_logger_wrapper.dart';
 import 'package:flutter_moshimoshi/entities/token.dart';
 
 class Tokens {